requirements.txt
aiohttp
aioredis 
aiomysql 
pyyaml 
sqlalchemy
click
aiocache